+++ static/js/translator/constants.ts
