+++ static/js/translator/page.tsx
@@ -18,82 +18,7 @@ import React from "react";
 
 import axios from "axios";
 
-interface Binding {
-  Mapping: {
-    Obj: string;
-    Pred: string;
-    Sub: string;
-  };
-  Query: {
-    Obj: string;
-    Pred: string;
-    Sub: string;
-  };
-}
-
-interface Constraint {
-  LHS: string;
-  RHS: string;
-}
-
-interface TranslationPropType {
-  sql: string;
-  bindings: Binding[];
-  constraints: Constraint[];
-}
-
-class Translation extends React.Component<TranslationPropType> {
-  render(): JSX.Element {
-    return (
-      <>
-        <div id="result">
-          <div className="title">Literal Matches</div>
-          {this.props.bindings.map((binding, index) => {
-            return (
-              <table className="literal-match-table" key={index}>
-                <tbody>
-                  <tr className="query-literal">
-                    <td>{binding.Mapping.Pred}</td>
-                    <td>{binding.Mapping.Sub}</td>
-                    <td>{binding.Mapping.Obj}</td>
-                  </tr>
-                  <tr>
-                    <td className="mapping-literal">{binding.Query.Pred}</td>
-                    <td className="mapping-literal">{binding.Query.Sub}</td>
-                    <td className="mapping-literal">{binding.Query.Obj}</td>
-                  </tr>
-                </tbody>
-              </table>
-            );
-          })}
-          <div className="title">Binding Sets and Translated SQL</div>
-          <table className="translation-info-table">
-            <tbody>
-              <tr className="header">
-                <td className="sql">SQL</td>
-                <td className="constraints">Constraints</td>
-              </tr>
-              <tr>
-                <td className="sql-result">{this.props.sql}</td>
-                <td>
-                  <ul>
-                    {this.props.constraints.map((c, index) => {
-                      return (
-                        <li key={index}>
-                          &#40;{c.LHS}&#44;&nbsp;{c.RHS}&#41;
-                        </li>
-                      );
-                    })}
-                  </ul>
-                </td>
-              </tr>
-            </tbody>
-          </table>
-        </div>
-      </>
-    );
-  }
-}
+import { Binding, Constraint, Translation } from "./translation";
 
 interface PagePropType {
   mapping: string;
