+++ static/js/translator/translator.ts
@@ -22,7 +22,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import { Page } from "./page";
-import { SCHEMA_MAPPING, SAMPLE_QUERY } from "./constant";
+import { SCHEMA_MAPPING, SAMPLE_QUERY } from "./constants";
 
 /**
  * Update translation results when schema mapping or query changes.
